requirements.txt
streamlit
transformers
torch==2.0.1
pdfplumber
python-docx
